Move registry fixture to conftest.
